refactor: 다리 생성 로직 BridgeGame 클래스로 이동
