+++ src/BridgeGame.js
@@ -1,13 +1,28 @@
+const BridgeMaker = require('./BridgeMaker');
+const BridgeRandomNumberGenerator = require('./BridgeRandomNumberGenerator');
+
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
 class BridgeGame {
+  #bridges;
+
+  /**
+   * @param {number} bridgeSize
+   * @param {number} numberOfRow
+   */
+  constructor(bridgeSize, numberOfRow) {
+    this.#bridges = this.makeBridges(bridgeSize, numberOfRow);
+  }
+
   /**
    * 사용자가 칸을 이동할 때 사용하는 메서드
    * <p>
    * 이동을 위해 필요한 메서드의 반환 값(return value), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
    */
-  move() {}
+  move() {
+    console.log('move');
+  }
 
   /**
    * 사용자가 게임을 다시 시도할 때 사용하는 메서드
@@ -15,6 +30,24 @@ class BridgeGame {
    * 재시작을 위해 필요한 메서드의 반환 값(return value), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
    */
   retry() {}
+
+  /**
+   * 여러줄의 다리를 생성하는 메서드
+   * @param {number} bridgeSize
+   * @param {number} numberOfRow
+   * @returns
+   */
+  makeBridges(bridgeSize, numberOfRow) {
+    const generate = BridgeRandomNumberGenerator.generate;
+
+    let bridges = [];
+    for (let i = 0; i < numberOfRow; i++) {
+      const bridge = BridgeMaker.makeBridge(bridgeSize, generate);
+      bridges.push(bridge);
+    }
+
+    return bridges;
+  }
 }
 
 module.exports = BridgeGame;
